+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - `mongodbreceiver`: Add integration test for mongodb receiver (#10864)
 - `mezmoexporter`: add logging for HTTP errors (#10875)
 - `googlecloudexporter`: Support writing to multiple GCP projects by setting the `gcp.project.id` resource attribute, and support service account impersonation (#11051)
+- `prometheusreceiver`: Add `target_info` labels to resource attributes. (#11034)
 - `k8sattributeprocessor`: Add debug logs to help identify missing attributes (#11060)
 - `jmxreceiver`: Add latest releases of jmx metrics gatherer & wildfly jar to supported jars hash list (#11134)
 - `rabbitmqreceiver`: Add integration test for rabbitmq receiver (#10865)